+++ example/next-tailwind-ts/src/app/page.tsx
@@ -1,0 +1,14 @@
+import DateInput from "@/component/DateInput";
+
+const Page = () => {
+  return (
+    <div className="rounded-lg bg-white px-12 py-8 shadow">
+      <div className="text-xl leading-loose font-bold">
+        thaidatepicker-react + NextJS Demo
+      </div>
+      <DateInput />
+    </div>
+  );
+};
+
+export default Page;