+++ src/utils/index.ts
@@ -1,3 +1,6 @@
+import { clsx, type ClassValue } from "clsx";
+import { twMerge } from "tailwind-merge";
+
 /**
  * Convert christ year to buddhist year (+543)
  */
@@ -13,3 +16,10 @@ export const GetHighlightByDate = (date = new Date()) => [
     "react-datepicker__day--highlighted-today": [date],
   },
 ];
+
+/**
+ * Merge and deduplicate classnames using tailwind-merge and clsx.
+ */
+export const cn = (...inputs: ClassValue[]) => {
+  return twMerge(clsx(inputs));
+};